+++ src/api/hooks/auth/login.api.ts
@@ -1,8 +1,8 @@
 import { useMutation } from '@tanstack/react-query';
 
-import type { UserRequestData, UserResponseData } from '@/types';
+import { BASE_URL, fetchInstance } from '@/api/instance';
 
-import { BASE_URL, fetchInstance } from '../instance';
+import type { AuthProps, UserRequestData, UserResponseData } from './type';
 
 export const getLoginPath = () => `${BASE_URL}/api/members/login`;
 
@@ -11,10 +11,8 @@ export const loginUser = async (userData: UserRequestData) => {
   return response.data;
 };
 
-type Props = UserRequestData;
-
 export const useLogin = () => {
   return useMutation({
-    mutationFn: (userData: Props) => loginUser(userData),
+    mutationFn: (userData: AuthProps) => loginUser(userData),
   });
 };
