+++ src/api/hooks/product/product-detail.api.ts
@@ -1,16 +1,12 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
 
-import type { ProductData } from '@/types';
+import { BASE_URL, fetchInstance } from '@/api/instance';
 
-import { BASE_URL, fetchInstance } from '../instance';
-
-export type ProductDetailRequestParams = {
-  productId: string;
-};
-
-type Props = ProductDetailRequestParams;
-
-export type GoodsDetailResponseData = ProductData;
+import type {
+  GoodsDetailResponseData,
+  ProductDetailProps,
+  ProductDetailRequestParams,
+} from './type';
 
 export const getProductDetailPath = (productId: string) => `${BASE_URL}/api/products/${productId}`;
 
@@ -22,7 +18,7 @@ export const getProductDetail = async (params: ProductDetailRequestParams) => {
   return response.data;
 };
 
-export const useGetProductDetail = ({ productId }: Props) => {
+export const useGetProductDetail = ({ productId }: ProductDetailProps) => {
   return useSuspenseQuery({
     queryKey: [getProductDetailPath(productId)],
     queryFn: () => getProductDetail({ productId }),
