+++ src/api/hooks/product/product-options.api.ts
@@ -1,13 +1,12 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
 
-import type { ProductOptionsData } from '@/types';
+import { BASE_URL, fetchInstance } from '@/api/instance';
 
-import { BASE_URL, fetchInstance } from '../instance';
-import type { ProductDetailRequestParams } from './useGetProductDetail';
-
-type Props = ProductDetailRequestParams;
-
-export type ProductOptionsResponseData = ProductOptionsData[];
+import type {
+  ProductDetailProps,
+  ProductDetailRequestParams,
+  ProductOptionsResponseData,
+} from './type';
 
 export const getProductOptionsPath = (productId: string) =>
   `${BASE_URL}/api/products/${productId}/options`;
@@ -19,7 +18,7 @@ export const getProductOptions = async (params: ProductDetailRequestParams) => {
   return response.data;
 };
 
-export const useGetProductOptions = ({ productId }: Props) => {
+export const useGetProductOptions = ({ productId }: ProductDetailProps) => {
   return useSuspenseQuery({
     queryKey: [getProductOptionsPath(productId)],
     queryFn: () => getProductOptions({ productId }),
