+++ src/api/hooks/wish-list/wish-list-find.api.ts
@@ -1,10 +1,10 @@
 import { useQuery } from '@tanstack/react-query';
 
-import type { ProductListResponse } from '@/types';
-
-import { BASE_URL, tokenInstance } from '../instance';
+import { BASE_URL, tokenInstance } from '@/api/instance';
+import type { ProductListResponse } from '@/types/product';
 
 export const getWishListPath = () => `${BASE_URL}/api/wishes?page=0&size=10&sort=createdDate`;
+
 const wishListQueryKey = [getWishListPath()];
 
 export const getWishList = async () => {
